Create highlightedTegahedron
This can use used to create a polygon my making a map from the red dot to the green equatorial points. The minimal distance from the red point to each successive equatorial points, along the cage (so more of a taxicab distance) could be done where the angle is the actual angle from the closest equatorial point to the successive equatorial point.

Also, interesting would be a billiard ball simulation where points can move through it like Bohr's electrons on a 'rail'. Rules could be changed to allow for the 'electrons' to pass by through each other, or they could have a distance that is allowed before repulsion and bouncing back.